build(docker): Update Dockerfile with comments and structure
Enhance the Dockerfile by adding comments to clarify the purpose of
each stage in the multi-stage build process. Set the base image to
`mcr.microsoft.com/dotnet/aspnet:9.0` and configure the user and
working directory. Expose ports 8080 and 8081. The build stage now
uses `mcr.microsoft.com/dotnet/sdk:9.0`, with project files copied
for restoration and building. The publish stage prepares the project
for deployment, and the final stage sets up the application for
production use, defining the entry point.